fix(textproto): handle tabs correctly
* unfortunately the proto tokenizer/parser counts tabs as 8 columns, so
  converting between column offset and byte offset is tricky. before
  this commit, the textproto analyzer failed to correctly index any lines
  of the file that included tabs
* moved proto indexer's ByteOffsetIntoLine() into a library which is
  shared with the textproto indexer
* added a test case that uses tabs
