+++ kythe/cxx/indexer/proto/file_descriptor_walker.cc
@@ -26,6 +26,7 @@
 #include "google/protobuf/stubs/map_util.h"
 #include "kythe/cxx/common/status_or.h"
 #include "kythe/cxx/indexer/proto/marked_source.h"
+#include "kythe/cxx/indexer/proto/offset_util.h"
 #include "kythe/cxx/indexer/proto/proto_graph_builder.h"
 #include "re2/re2.h"
 #include "re2/stringpiece.h"
@@ -81,30 +82,6 @@ class ScopedLookup {
   std::vector<int>* lookup_path_;
   const int component_;
 };
-
-// Figures out just how many bytes one needs to go into `line_text` to reach
-// what the proto compiler calls column `column_number`.
-int ByteOffsetIntoLine(int column_number, absl::string_view line_text) {
-  int computed_column = 0;
-  int offset = 0;
-  while (computed_column < column_number && offset < line_text.size()) {
-    if (line_text[offset] == '\t') {
-      // In proto land, tabs go to the next multiple of 8.  There are a million
-      // ways of computing this.  This one will do.
-      computed_column = (computed_column + 8) - (computed_column % 8);
-    } else {
-      ++computed_column;
-    }
-    ++offset;
-  }
-  if (computed_column != column_number) {
-    LOG(ERROR) << "Error computing byte offset: expected " << column_number
-               << " columns but counted up to " << computed_column
-               << " in line \"" << line_text << "\"";
-    return -1;
-  }
-  return offset;
-}
 
 }  // namespace
